requirements.txt
asgiref==3.6.0
colorama==0.4.6
crispy-bootstrap5==0.7
Django==3.2
django-crispy-forms==2.0
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
exceptiongroup==1.1.1
flake8==6.0.0
iniconfig==2.0.0
mccabe==0.7.0
packaging==23.1
phonenumberslite==8.13.11
Pillow==9.5.0
pluggy==1.0.0
pycodestyle==2.10.0
pyflakes==3.0.1
pytest==7.3.1
pytest-django==4.5.2
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
tomli==2.0.1
